32GSMgatewayServer/requirements.txt
asgiref==3.7.2
asterisk-ami==0.1.7
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.6
numpy==1.26.3
pandas==2.1.4
psutil==5.9.7
pyst2==0.5.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
subprocess.run==0.0.8
tzdata==2023.4
urllib3==2.1.0
asgiref==3.7.2
asterisk-ami==0.1.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.1
django-rest-framework==0.1.0
django-sslserver==0.22
djangorestframework==3.14.0
gTTS==2.5.1
idna==3.6
numpy==1.26.3
pandas==2.1.4
pexpect==4.9.0
psutil==5.9.7
ptyprocess==0.7.0
pydub==0.25.1
pyst2==0.5.1
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
subprocess.run==0.0.8
tzdata==2023.4
urllib3==2.1.0
PyJWT>=2.8.0
cryptography>=42.0.5
